refactor(web): move auth sign-in page into features/auth
Preserve history with git mv. Wrap SSOButtons empty branch so the file
passes @repo/no-null-render under features/. Route shim lands next.
